move lern-store components to lern-store folder
