+++ src/components/lern-store/LernstoreCollectionDetailView.vue
@@ -138,9 +138,9 @@ import { printDateFromTimestamp } from "@/plugins/datetime";
 import contentMeta from "@/mixins/contentMeta";
 import infiniteScrolling from "@/mixins/infiniteScrolling";
 import { contentModule, notifierModule } from "@/store";
-import AddContentButton from "@/components/organisms/AddContentButton";
-import ContentCard from "@/components/organisms/ContentCard";
-import ContentEduSharingFooter from "@/components/molecules/ContentEduSharingFooter";
+import AddContentButton from "@/components/lern-store/AddContentButton";
+import ContentCard from "@/components/lern-store/ContentCard";
+import ContentEduSharingFooter from "@/components/lern-store/ContentEduSharingFooter";
 import UserHasRole from "@/components/helpers/UserHasRole";
 import BaseLink from "../base/BaseLink";
 import { RenderHTML } from "@feature-render-html";
