+++ src/components/lern-store/LernstoreDetailView.vue
@@ -175,7 +175,7 @@
 
 <script>
 /* eslint-disable max-lines */
-import AddContentButton from "@/components/organisms/AddContentButton";
+import AddContentButton from "@/components/lern-store/AddContentButton";
 import UserHasRole from "@/components/helpers/UserHasRole";
 import contentMeta from "@/mixins/contentMeta";
 import LernStorePlayer from "@/components/lern-store/LernStorePlayer";
